+++ web/app/routes/changelog.tsx
@@ -3,22 +3,26 @@ import type { Route } from "./+types/changelog";
 import { CHANGELOG_META_FUNCTION } from "app/meta/changelog";
 
 export interface ILoaderData_Changelog {
-  data: string;
+  data: string | null;
+  error: boolean;
 }
 
 // Get changelog info
 export async function loader() {
-  const envUrl =
-    process.env.HOST === "https://khofly.com"
-      ? "https://raw.githubusercontent.com/cufta22/khofly/refs/heads/master/CHANGELOG.md"
-      : "https://raw.githubusercontent.com/cufta22/khofly/refs/heads/staging/CHANGELOG.md";
-  const a = "https://raw.githubusercontent.com/cufta22/khofly/refs/heads/staging/instances.json";
+  try {
+    const envUrl =
+      process.env.HOST === "https://khofly.com"
+        ? "https://raw.githubusercontent.com/cufta22/khofly/refs/heads/master/CHANGELOG.md"
+        : "https://raw.githubusercontent.com/cufta22/khofly/refs/heads/staging/CHANGELOG.md";
 
-  const data = await fetch(envUrl);
+    const data = await fetch(envUrl);
 
-  const changelog = await data.text();
+    const changelog = await data.text();
 
-  return { data: changelog };
+    return { data: changelog, error: false };
+  } catch (error) {
+    return { data: null, error: true };
+  }
 }
 
 // Meta tags