+++ web/src/modules/Instances/index.tsx
@@ -40,9 +40,20 @@ const getCspColor = (val: string) => {
 const PageInstances: React.FC<Props> = ({ loaderData }) => {
   const t = useTranslate();
 
+  console.log(loaderData);
+
   const linkTextColor = usePrimaryColor(4);
 
-  if (!loaderData.data)
+  // Rip API
+  if (loaderData.error)
+    return (
+      <Center mt={100}>
+        <Text size="xl">Rip API :(</Text>
+      </Center>
+    );
+
+  // Data is loading
+  if (!loaderData?.data)
     return (
       <Center mt={100}>
         <Loader size="xl" />